do usr : testing init process

+++ include/dev/intrs.h
@@ -13,7 +13,7 @@ void irq_mask(uint8_t irq_num, bool set);
 
 void intrs_setup(void);
 
-void irq_set_handler(uint32_t irq_num, intr_handler_f handler);
+void irq_set_handler(uint8_t irq_num, intr_handler_f handler);
 
 void * intr_stack_ret_addr(void);
